+++ freshbooks/builders/paginator.py
@@ -111,6 +111,10 @@ class PaginateBuilder(Builder):
     def build(self, resource_name: Optional[str] = None) -> str:
         """Builds the query string parameters from the PaginateBuilder.
 
+        Args:
+            resource_name: The type of resource to generate the query string for.
+                           Eg. AccountingResource, ProjectsResource
+
         Returns:
             The built query string
         """